requirements.txt
discord.py
pymongo
colorama
python-dotenv